Use locally built cbioportal
Was only able to figure out how to use core. That lacked some deps such
as log4j and mysql. A quick hack was to just include the log4j dep from
cbioportal and remove the references to the mysql package. Being able to
use the locally built cbioportal will be useful to do local debugging
without needing jitpack.